+++ CrySPY/interface/select_code.py
@@ -4,6 +4,8 @@
 from VASP import calc_files_vasp, ctrl_job_vasp, collect_vasp
 from QE import calc_files_qe, ctrl_job_qe, collect_qe
 from soiap import calc_files_soiap, ctrl_job_soiap, collect_soiap
+from LAMMPS import calc_files_lammps, ctrl_job_lammps, collect_lammps
+
 from ..IO import read_input as rin
 
 
@@ -14,8 +16,10 @@ def check_calc_files():
         calc_files_qe.check_input_qe()
     elif rin.calc_code == 'soiap':
         calc_files_soiap.check_input_soiap()
+    elif rin.calc_code == 'LAMMPS':
+        calc_files_lammps.check_input_lammps()
     else:
-        raise SystemExit('now only VASP, QE, or soiap')
+        raise NotImplementedError('now only VASP, QE, soiap, or LAMMPS')
 
 
 def next_stage(stage, work_path, *args):
@@ -30,8 +34,11 @@ def next_stage(stage, work_path, *args):
     elif rin.calc_code == 'soiap':
         skip_flag = ctrl_job_soiap.next_stage_soiap(stage, work_path)
         return skip_flag
+    elif rin.calc_code == 'LAMMPS':
+        skip_flag = ctrl_job_lammps.next_stage_lammps(stage, work_path)
+        return skip_flag
     else:
-        raise SystemExit('now only VASP, QE, or soiap')
+        raise NotImplementedError('now only VASP, QE, soiap, or LAMMPS')
 
 
 def collect(current_id, work_path, check_file=None):
@@ -44,8 +51,11 @@ def collect(current_id, work_path, check_file=None):
     elif rin.calc_code == 'soiap':
         opt_struc, energy, magmom, check_opt = \
             collect_soiap.collect_soiap(current_id, work_path)
+    elif rin.calc_code == 'LAMMPS':
+        opt_struc, energy, magmom, check_opt = \
+            collect_lammps.collect_lammps(current_id, work_path)
     else:
-        raise SystemExit('now only VASP, QE, or soiap')
+        raise NotImplementedError('now only VASP, QE, soiap, or LAMMPS')
 
     # ---------- return
     return opt_struc, energy, magmom, check_opt
@@ -61,8 +71,10 @@ def next_struc(structure, next_id, work_path, *args):
         return kpt_data
     elif rin.calc_code == 'soiap':
         ctrl_job_soiap.next_struc_soiap(structure, next_id, work_path)
+    elif rin.calc_code == 'LAMMPS':
+        ctrl_job_lammps.next_struc_lammps(structure, next_id, work_path)
     else:
-        raise SystemExit('now only VASP, QE, or soiap')
+        raise NotImplementedError('now only VASP, QE, soiap, or LAMMPS')
 
 
 def clean_calc_files(work_path):
@@ -72,8 +84,10 @@ def clean_calc_files(work_path):
         calc_files_qe.clean_calc_files_qe(work_path)
     elif rin.calc_code == 'soiap':
         calc_files_soiap.clean_calc_files_soiap(work_path)
+    elif rin.calc_code == 'LAMMPS':
+        calc_files_lammps.clean_calc_files_lammps(work_path)
     else:
-        raise SystemExit('now only VASP, QE, or soiap')
+        raise NotImplementedError('now only VASP, QE, soiap, or LAMMPS')
 
 
 def get_energy_step(energy_step_data, current_id, work_path):
@@ -81,11 +95,13 @@ def get_energy_step(energy_step_data, current_id, work_path):
         energy_step_data = collect_vasp.get_energy_step_vasp(energy_step_data, current_id, work_path+'vasprun.xml')
         return energy_step_data
     elif rin.calc_code == 'QE':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
     elif rin.calc_code == 'soiap':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
+    elif rin.calc_code == 'LAMMPS':
+        raise NotImplementedError('now only VASP')
     else:
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
 
 
 def get_struc_step(struc_step_data, current_id, work_path):
@@ -93,11 +109,13 @@ def get_struc_step(struc_step_data, current_id, work_path):
         struc_step_data = collect_vasp.get_struc_step_vasp(struc_step_data, current_id, work_path+'vasprun.xml')
         return struc_step_data
     elif rin.calc_code == 'QE':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
     elif rin.calc_code == 'soiap':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
+    elif rin.calc_code == 'LAMMPS':
+        raise NotImplementedError('now only VASP')
     else:
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
 
 
 def get_fs_step(fs_step_data, current_id, work_path):
@@ -105,8 +123,10 @@ def get_fs_step(fs_step_data, current_id, work_path):
         fs_step_data = collect_vasp.get_fs_step_vasp(fs_step_data, current_id, work_path+'vasprun.xml')
         return fs_step_data
     elif rin.calc_code == 'QE':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
     elif rin.calc_code == 'soiap':
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')
+    elif rin.calc_code == 'LAMMPS':
+        raise NotImplementedError('now only VASP')
     else:
-        raise SystemExit('now only VASP')
+        raise NotImplementedError('now only VASP')