+++ example/0.3.0_QE_RS_Si8_asahi/cryspy.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-python /home/yamashita06/CrySPY/0.3.0/cryspy.py 1>> log
-
-