+++ utility/spg_check.py
@@ -4,18 +4,17 @@
 import argparse
 
 from pymatgen import Structure
-from pymatgen.io.cif import CifWriter
 
 
-def get_cif(filename, tolerance=0.1):
+def get_spg_info(filename, tolerance=0.1):
     if filename[-5:] == '.vasp':
         with open(filename, 'r') as f:
             struc_str = f.read()
         struc = Structure.from_str(struc_str, fmt='poscar')
     else:
         struc = Structure.from_file(filename)
-    cif = CifWriter(struc, symprec=tolerance)
-    cif.write_file(filename+'.cif')
+    spg_sym, spg_num = struc.get_space_group_info(symprec=tolerance)
+    return spg_sym, spg_num
 
 
 if __name__ == '__main__':
@@ -27,4 +26,4 @@ if __name__ == '__main__':
 
     # ---------- main
     filename = args.infile.split('/')[-1]    # ./aaa/bbb/POSCAR --> POSCAR
-    get_cif(filename, args.tolerance)
+    print get_spg_info(filename, args.tolerance)