+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.6.2] - 2018-06-26
+### Added
+- LAMMPS can be used in CrySPY
+
 ## [0.6.1] - 2018-03-01
 ### Fixed
 - Fixed a bug for job control in RS
@@ -9,7 +13,7 @@
 - LAQA is now available
 
 ### Changed
-- Changed the data format of init_struc_data and opt_struc_data from int to dict
+- Changed the data format of init_struc_data and opt_struc_data from list to dict
 
 ### Fixed
 - Fixed a minor bug for BO